@@ -15,42 +15,39 @@
     }
   ],
   "require": {
-    "php": "^7.1.3|^8.0",
-    "latte/latte": "^2.5",
-    "nette/di": "^3.0",
-    "nette/finder": "^2.5",
+    "php": "^7.4|^8.0",
+    "latte/latte": "^2.6",
+    "nette/di": "^3.0.6",
+    "nette/finder": "^2.5.2",
     "nette/http": "^3.0",
     "nette/neon": "^3.0",
     "nette/schema": "^1.0",
     "nette/routing": "^3.0",
-    "nette/utils": "^3.0",
-    "symfony/translation": "^4.4|^5.0",
-    "symfony/config": "^4.4|^5.0"
+    "nette/utils": "^3.2.1",
+    "symfony/translation": "^5.0",
+    "symfony/config": "^5.0"
   },
   "require-dev": {
-    "doctrine/orm": "^2.6",
-    "mockery/mockery": "^1.2",
+    "doctrine/orm": "^2.8",
+    "mockery/mockery": "^1.4",
     "nette/application": "^3.0",
     "nette/bootstrap": "^3.0",
-    "nette/database": "^3.0",
-    "nette/robot-loader": " ^3.0",
+    "nette/database": "^3.1.1",
+    "nette/robot-loader": " ^3.2",
     "nette/tester": "^2.3.1",
     "ninjify/nunjuck": "^0.3.0",
     "ninjify/qa": "^0.12",
     "phpstan/phpstan": "^0.12.2",
-    "phpstan/phpstan-deprecation-rules": "^0.12.0",
-    "phpstan/phpstan-nette": "^0.12.0",
+    "phpstan/phpstan-deprecation-rules": "^0.12",
+    "phpstan/phpstan-nette": "^0.12",
     "phpstan/phpstan-strict-rules": "^0.12.5",
-    "psr/log": "^1.0",
-    "tracy/tracy": "^2.6"
+    "psr/log": "^1.1",
+    "tracy/tracy": "^2.8"
   },
   "autoload": {
     "psr-4": {
       "Contributte\\Translation\\": "src"
-    },
-    "classmap": [
-      "src/compatibility.php"
-    ]
+    }
   },
   "autoload-dev": {
     "psr-4": {
@@ -62,7 +59,7 @@
   },
   "extra": {
     "branch-alias": {
-      "dev-master": "0.9.x-dev"
+      "dev-master": "1.0.x-dev"
     }
   }
 }
