@@ -15,41 +15,33 @@ use Symfony\Component\Translation\Translator as SymfonyTranslator;
 class Translator extends SymfonyTranslator implements ITranslator
 {
 
-	/** @var \Contributte\Translation\LocaleResolver */
-	private $localeResolver;
+	private LocaleResolver $localeResolver;
 
-	/** @var \Contributte\Translation\FallbackResolver */
-	private $fallbackResolver;
+	private FallbackResolver $fallbackResolver;
 
-	/** @var string */
-	private $defaultLocale;
+	private string $defaultLocale;
 
-	/** @var string|null */
-	private $cacheDir;
+	private ?string $cacheDir;
 
-	/** @var bool */
-	private $debug;
+	private bool $debug;
 
-	/** @var bool */
-	public $returnOriginalMessage = true;
+	public bool $returnOriginalMessage = true;
 
 	/** @var array<string>|null */
-	private $localesWhitelist;
+	private ?array $localesWhitelist = null;
 
 	/** @var array<string> */
-	private $prefix = [];
+	private array $prefix = [];
 
 	/** @var array<array<string>> */
-	private $prefixTemp = [];
+	private array $prefixTemp = [];
 
 	/** @var array<bool> */
-	private $resourcesLocales = [];
+	private array $resourcesLocales = [];
 
-	/** @var \Psr\Log\LoggerInterface|null */
-	private $psrLogger;
+	private ?LoggerInterface $psrLogger = null;
 
-	/** @var \Contributte\Translation\Tracy\Panel|null */
-	private $tracyPanel;
+	private ?Panel $tracyPanel = null;
 
 	/** @var string */
 	private $initLang;
@@ -208,27 +200,18 @@ class Translator extends SymfonyTranslator implements ITranslator
 		return $locales;
 	}
 
-	/**
-	 * @param string $format
-	 * @param mixed $resource
-	 * @param string $locale
-	 * @param string|null $domain
-	 */
 	public function addResource(
-		$format,
+		string $format,
 		$resource,
-		$locale,
-		$domain = null
-	)
+		string $locale,
+		?string $domain = null
+	): void
 	{
 		parent::addResource($format, $resource, $locale, $domain);
 		$this->resourcesLocales[$locale] = true;
 	}
 
-	/**
-	 * {@inheritdoc}
-	 */
-	public function getLocale()
+	public function getLocale(): string
 	{
 		if (parent::getLocale() === $this->initLang) {
 			$this->setLocale($this->localeResolver->resolve($this));
@@ -238,11 +221,11 @@ class Translator extends SymfonyTranslator implements ITranslator
 	}
 
 	/**
-	 * {@inheritdoc}
+	 * @param array<string> $locales
 	 */
 	public function setFallbackLocales(
 		array $locales
-	)
+	): void
 	{
 		parent::setFallbackLocales($locales);
 		$this->fallbackResolver->setFallbackLocales($locales);
@@ -351,18 +334,14 @@ class Translator extends SymfonyTranslator implements ITranslator
 	}
 
 	/**
-	 * @param string|null $id
 	 * @param array<mixed> $parameters
-	 * @param string|null $domain
-	 * @param string|null $locale
-	 * @return string
 	 */
 	public function trans(
-		$id,
+		?string $id,
 		array $parameters = [],
-		$domain = null,
-		$locale = null
-	)
+		?string $domain = null,
+		?string $locale = null
+	): string
 	{
 		if ($domain === null) {
 			$domain = 'messages';
@@ -390,11 +369,11 @@ class Translator extends SymfonyTranslator implements ITranslator
 	}
 
 	/**
-	 * {@inheritdoc}
+	 * @return array<string>
 	 */
 	protected function computeFallbackLocales(
-		$locale
-	)
+		string $locale
+	): array
 	{
 		return $this->fallbackResolver->compute($this, $locale);
 	}
